Question_Complex_form.html
This HTML code creates a Complex Form that is divided into sections for Personal Information, Address, and Other Information. It gathers details such as name, email, phone, date of birth, gender, country, address, and extra comments. There is also a checkbox for agreeing to the terms. A Submit button allows users to send the data.