Plain-Function and Arrow-Function.js
Demonstrates the use of functions and loops. It defines an arrow function getExp(x, y) that calculates and returns the exponent value of x raised to y. Then it prints numbers from 1 to 15 using three types of loops  a for loop, a while loop, and a do-while loop  showing different ways to iterate through numbers.